kdelcjhai

+++ main.cpp
@@ -4,7 +4,7 @@ int f = 0;
   std::cout << "F|C" << std::endl;
   std::cout << "---" << std::endl;
   for (; f < 21;){
-    double c = (5.0/9.0) * (f-32);
+    int c = static_cast<int>((5.0/9.0) * (f-32));
     std::cout << f << "|" << c << std::endl;
     f++;
   }
